+++ views/frontpage.handlebars
@@ -32,9 +32,9 @@
 	<div class="container homesearch">
 		<div class="row">
 			<div class="col-4 offset-4">
-				<form class="clearfix mb-4" action="GET">
+				<form class="clearfix mb-4" method= "get" action="searchresults">
 					<div class="form-group">
-						<input type="text" id="search" class="form-control shadow-lg p-4" aria-describedby="search" placeholder="Zip code?">
+						<input type="text" name="zip" type="zip" id="zip" class="form-control shadow-lg p-4" aria-describedby="search" placeholder="Zip code?">
 					</div>
 					<button id="searchbutton" class="btn btn-danger btn-lg shadow-lg searchbutton" data-toggle="button" aria-pressed="false" autocomplete="off" role="button" aria-pressed="true">Let's get munching!</button>
 				</form>
@@ -50,7 +50,6 @@
     </div>
   </div>
 </nav>
-
 
 <style>
 	body {
