+++ views/layouts/main.handlebars
@@ -21,52 +21,9 @@
 
 {{{body}}}
 
-<script src="https://ajax.googleapis.com/ajax/libs/jquery/3.3.1/jquery.min.js"></script>
 
 <script src="https://stackpath.bootstrapcdn.com/bootstrap/4.1.3/js/bootstrap.min.js" integrity="sha384-ChfqqxuZUCnJSK3+MXmPNIyE6ZbWh2IMqE241rYiqJxyMiZ6OW/JmZQ5stwEULTy" crossorigin="anonymous"></script>
 </body>
-
-<script type="text/javascript">
-
-
-  $(".searchbutton").click(function(){
-    event.preventDefault(); 
-    var url = new URL(window.location);
-    var location = $("#search").val();
-    url.searchParams.set("userID", "2");
-    console.log(url);
-    window.location='/searchresults/' + location + url.search;
-    
-  })
-
-   $(".details").click(function(){
-    event.preventDefault(); 
-    var url = new URL(window.location);
-    var urlParam = url.search;
-    var favoriteID = $(this).attr("id");
-    var favoriteName = $(this).attr("name");
-    var favoriteRating = $(this).attr("rating");
-    var favoritePhone = $(this).attr("phone");
-
-    var query = new URLSearchParams(urlParam);
-    var user = query.get("userID");
-    
-    var dataObject = {
-      UserinfoId: user,
-      favoritesID: favoriteID
-    }
-
-    console.log(dataObject);
-
-    $.ajax({
-      type: "POST",
-      url: "/api/favorites",
-      data: dataObject,
-    });
-
-  })
-
-</script>
 
 
 </html>
