+++ views/searchresults.handlebars
@@ -71,6 +71,9 @@
   </div>
 </nav>
 
+<script type="text/javascript" src="https://cdnjs.cloudflare.com/ajax/libs/jquery/3.2.1/jquery.min.js"></script>
+<script src="../public/js/index.js"></script>
+
 
 <style>
 	body {
